requirements.txt
blinker==1.9.0
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
Flask==2.2.5
flask-mongoengine==1.0.0
Flask-WTF==1.2.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.3.0
marshmallow-mongoengine==0.31.2
mongoengine==0.29.3
PyJWT==2.12.1
pymongo==4.16.0
python-dotenv==1.2.2
Werkzeug==3.1.8
WTForms==3.2.1
